gen-0/requirements.txt
aiohttp>=3.9
pytest-asyncio>=0.24
